# internal/dto/dto.go
package dto

type CreateProductInput struct {
	Name  string  `json:"name"`
	Price float64 `json:"price"`
}

type CreateUserInpt struct {
	Name     string `json:"name"`
	Email    string `json:"email"`
	Password string `json:"password"`
}

type GenerateTokenInput struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type GenerateTokenOutput struct {
	AccessToken string `json:"access_token"`
}